Fix things, warnings and remove stuff

+++ nine_morris_3d_engine/src/application_base/window.cpp
@@ -193,7 +193,7 @@ namespace sm {
             glfw_icons.push_back(image);
         }
 
-        glfwSetWindowIcon(window, glfw_icons.size(), glfw_icons.data());
+        glfwSetWindowIcon(window, static_cast<int>(glfw_icons.size()), glfw_icons.data());
     }
 
     double Window::get_time() {
